feat(global): add Metamask network switch to Ethereum mainnet
If Metamask is connected to a different network than Ethereum, and the 'from'
token resides on Ethereum, the 'Connect Wallet' operation does nothing.
This commit addresses that issue, by introducing a new function changeNetwork()
which internally calls the RPC method wallet_switchEthereumChain:
in this way, Metamask can correctly switch to Ethereum.