feat(store): revise opt-in procedure for pTokens on Algorand
Now, dapp checks if the Algorand destination account has opted-in the destination
pToken. If the connected Algorand wallet account matches the specified
destination account, then a request for a transaction signature is sent
to the wallet to opt-in.
In case the destination account hasn't opted-in, the Swap button is not
clickable and notifying that.