fix(store): revise USDT allowance setting via approve method
That dapp calls the approval for any and all EVM peg-ins, usually of the peg-in amount.
This is because the pegIn function works by users having to call it, rather than users
just sending tokens. As such, the context the pegIn fxn is called from is the point of view
of the contract, which calls transferFrom() in order to move the tokens from the pegIn caller
to itself. But in order to do that, the user must first have allowed the contract
to spend X tokens on their behalf. Which is what approve is for. And which is why
the DApp calls that first.

Note that not all contracts are using the attack mitigation above that requires the balance
to first be set to zero before changing the approval. If there is sufficient allowance,
the approval call should be skipped, since it's a waste of funds. If allowance is zero,
call to approve as normal.

If allowance is > zero but < than required, that's where it gets tricky. If you can get
a gas estimate for the approve call, that'll help figure out if this attack mitigation is in place.
If that's the case, you'll have to make the approve call twice, first for 0 then for
the required amount. If it's not the case, just call the approve with the required amount.

Credits: Greg Kapka